requirements.txt
asgiref==3.7.2
Django==5.0.2
django-etc==1.4.0
django-hitcount==1.3.5
django-modeltranslation==0.18.11
marshmallow==3.21.1
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.9
python-decouple==3.8
python-dotenv==1.0.1
sqlparse==0.4.4
typing_extensions==4.10.0
whitenoise==6.6.0
